+++ vaadin-testbench-it/src/test/java/com/vaadin/testbench/tests/ui/element/NativeButtonElement.java
@@ -1,7 +1,7 @@
 package com.vaadin.testbench.tests.ui.element;
 
 import com.vaadin.testbench.TestBenchElement;
-import com.vaadin.testbench.annotations.Element;
+import com.vaadin.testbench.elementsbase.Element;
 
 @Element("button")
 public class NativeButtonElement extends TestBenchElement {
