+++ vaadin-testbench-it/src/test/java/com/vaadin/testbench/tests/ui/element/PolymerTemplateViewElement.java
@@ -1,7 +1,7 @@
 package com.vaadin.testbench.tests.ui.element;
 
 import com.vaadin.testbench.TestBenchElement;
-import com.vaadin.testbench.annotations.Element;
+import com.vaadin.testbench.elementsbase.Element;
 
 @Element("polymer-template-view")
 public class PolymerTemplateViewElement extends TestBenchElement {
